[#35] Fixed up spotlight images for different screen sizes.
Have 2 images - one on top, one on the bottom of the article. The top is shown on small screens and the bottom is shown on large screens.

On large screens, the technique used in http://www.pagecolumn.com/liquidfixed/2_col_fix_liquid.htm is used. Both items on the row are floated left, and the fixed width frame has a negative 100% margin.